Conflicts:
	src/java/org/jsfml/window/Window.java
	src/java/org/jsfml/window/event/Event.java
	src/java/org/jsfml/window/event/JoystickButtonEvent.java
	src/java/org/jsfml/window/event/JoystickConnectEvent.java
	src/java/org/jsfml/window/event/JoystickEvent.java
	src/java/org/jsfml/window/event/JoystickMoveEvent.java
	src/java/org/jsfml/window/event/KeyEvent.java
	src/java/org/jsfml/window/event/MouseButtonEvent.java
	src/java/org/jsfml/window/event/MouseEvent.java
	src/java/org/jsfml/window/event/MouseWheelEvent.java
	src/java/org/jsfml/window/event/SizeEvent.java
	src/java/org/jsfml/window/event/TextEvent.java

Fixes #41

+++ src/java/org/jsfml/graphics/AWTRenderWindow.java
@@ -64,7 +64,7 @@ final class AWTRenderWindow extends RenderWindow implements
 
     @Override
     public void keyTyped(KeyEvent e) {
-        queue.add(new org.jsfml.window.event.TextEvent(Event.Type.TEXT_ENTERED.ordinal(), (long) e.getKeyChar()));
+        queue.add(new org.jsfml.window.event.TextEvent(Event.Type.TEXT_ENTERED.ordinal(), e.getKeyChar()));
     }
 
     @Override
